Moved each model to own file
