+++ Models/QuizResultsViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Sergey Simashov on 28.05.2024.
+//
+
+import Foundation
+
+private struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+}
