+++ Models/QuizStepViewModel.swift
@@ -1,0 +1,14 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Sergey Simashov on 28.05.2024.
+//
+
+import UIKit
+
+private struct QuizStepViewModel{
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+}
